move internal hooks into hooks dir
