+++ packages/react/src/components/RoomName.tsx
@@ -3,7 +3,7 @@ import type { Room } from 'livekit-client';
 import * as React from 'react';
 
 import { useEnsureRoom } from '../context';
-import { useObservableState } from '../helper/useObservableState';
+import { useObservableState } from '../hooks/internal/useObservableState';
 
 export interface UseRoomInfoOptions {
   room?: Room;
