+++ packages/react/src/hooks/internal/resizeObserver.ts
@@ -6,6 +6,8 @@ import useLatest from '@react-hook/latest';
 /**
  * A React hook that fires a callback whenever ResizeObserver detects a change to its size
  * code extracted from https://github.com/jaredLunde/react-hook/blob/master/packages/resize-observer/src/index.tsx in order to not include the polyfill for resize-observer
+ *
+ * @internal
  */
 export function useResizeObserver<T extends HTMLElement>(
   target: React.RefObject<T>,
