+++ packages/react/src/hooks/useIsSpeaking.ts
@@ -2,7 +2,7 @@ import { createIsSpeakingObserver } from '@livekit/components-core';
 import { Participant } from 'livekit-client';
 import * as React from 'react';
 import { useEnsureParticipant } from '../context';
-import { useObservableState } from '../helper';
+import { useObservableState } from './internal';
 
 export function useIsSpeaking(participant?: Participant) {
   const p = useEnsureParticipant(participant);
