+++ packages/react/src/hooks/useLocalParticipantPermissions.ts
@@ -1,7 +1,7 @@
 import { participantPermissionObserver } from '@livekit/components-core';
 import * as React from 'react';
 import { useRoomContext } from '../context';
-import { useObservableState } from '../helper';
+import { useObservableState } from './internal';
 
 export function useLocalParticipantPermissions() {
   const room = useRoomContext();
