+++ packages/react/src/hooks/useRemoteParticipant.ts
@@ -2,7 +2,7 @@ import { connectedParticipantObserver } from '@livekit/components-core';
 import { ParticipantEvent, RemoteParticipant } from 'livekit-client';
 import * as React from 'react';
 import { useRoomContext } from '../context';
-import { useObservableState } from '../helper';
+import { useObservableState } from './internal';
 
 export interface UseRemoteParticipantOptions {
   updateOnlyOn?: ParticipantEvent[];
